[BEAM-6186] Finish moving optimization phases.
This is mostly a simple refactoring, moving references by closure to
an explicit context argument.  Of note, however, the length-prefix coder
computation was reworked to be cleaner and more consistent.
